Fix New Year observance for Saturdays.
https://www.nyse.com/publicdocs/nyse/regulation/nyse/NYSE_Rules.pdf
Rule 7.2.

When a holiday observed by the Exchange falls on a Saturday, the Exchange will not be open for business on
the preceding Friday and when any holiday observed by the Exchange falls on a Sunday, the Exchange will not
be open for business on the succeeding Monday, unless unusual business conditions exist, such as the ending
of a monthly or yearly accounting period.