Split out extra packages
Split out three dependencies that are only used for some specific
countries so they're not installed by default anymore. This is a
breaking change for consumers using one or more of the affected
countries. The advantage is that consumers who only use one or a few
simple calendars don't need to install dependencies that will be unused.

This fixes #362.