Use common abbreviations for Austrian states.
Commit
https://github.com/vacanza/python-holidays/commit/4342d28ff0f6d6a63aff433c72b80d33589991f7
uses the numbers 1-9 for the nine states in Austria, but actually no one
in Austria uses these numbers. This commit uses the common character
codes (B, K, N, O, S, St, T, V, W) with the same default (9 and W) and
adds the numbers as deprecated subdivisions. (See also #248)

It is expected for the impact to be negligible, because none of these
subdivisions actually does have an effect on the list of holidays.