Adds aliases for subdivisions.
Aliases for subdivisions can be added in the `subdivisions_aliases`
class attribute, a dictionary with the alias as key and the ISO3166-2
subdivision code as value. More than one alias can be added by adding
more entries into the dictionary.

The alias will be resolved only when populating the holidays for the
subregion.