Add Finnish unofficial holidays
Current implementation generates only list of Public holidays (so
holidays established by Parliament). However there are also quite a lot
of so called "flag flying days" ("liputuspäivät" in Finnish or "flaggdagar"
in Swedish) that don't have official holiday status, but still it's
quite common for businesses to have shorter opening hours or to be
completely closed.

So this adds a list of those holidays under unofficial category.
The English and Swedish translations are taken from the
Ministry of the Interior website, and for Ukrainian one I asked a
Ukrainian friend of mine (who's native language is Ukrainian) to
review the list for me.

Sources:

- https://en.wikipedia.org/wiki/Flag_flying_days_in_Finland
- https://intermin.fi/en/flag-and-arms/flag-flying-days (Ministry of the
  Interior of Finland)